--- client/src/dummy/mypage.ts
@@ -1,19 +1,0 @@
-export const dummyBio = [
-  {
-    nickname: 'Taewoo Kim',
-    email: 'taewoo123@naver.com',
-    phoneNumber: '010-1234-5678',
-    profileimg:
-      'https://i.pinimg.com/564x/62/a6/1c/62a61c78a2228e23c14fb5b27951c5df.jpg',
-    stores: [
-      {
-        businessCategory: '해상 오락관련 서비스업',
-        storeName: '레저 좋아',
-      },
-      {
-        businessCategory: '육상 오락관련 서비스업',
-        storeName: '태우네 승마장',
-      },
-    ],
-  },
-];
